+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed:
+- Margin below habit completion dialog
+- Awaiting in getIt init
+
+### Changed:
+- Upgraded dependencies
+
+## [0.8.341] - 2023-05-02
 ### Changed:
 - Upgraded dependencies
 - Show config flag for invalid cert